Refactor backends into seperate files and added tests for extlinux
